Comparação de tempo e integração com motor (teste)
